api/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0